requirements.txt

openai==1.12.0
anthropic==0.18.1

transformers==4.37.2
textblob==0.17.1

pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1

textblob==0.17.1

numpy==1.26.3
scikit-learn==1.4.0

tenacity==8.2.3

click==8.1.7

matplotlib==3.8.2
seaborn==0.13.1
pandas==2.2.0

flask>=3.0.0
flask-cors>=4.0.0